+++ backend/run.py
@@ -3,48 +3,52 @@ import asyncio
 import subprocess
 from watchgod import awatch
 
+import asyncio
+import os
+import subprocess
+
 class ServerManager:
     def __init__(self):
-        print('init called')
-        print(os.getpid())
+        print('init called', os.getpid())
         self.process = None
 
     async def run_server(self):
-        print('running called')
-        print(os.getpid())
         if self.process:
-            print('ruin called')
+            print('terminating process')
             print(self.process, self.process.poll()) 
             self.process.terminate()
+            # Wait for the process to terminate
+            self.process.wait()
             await asyncio.sleep(1) 
-        self.process = subprocess.Popen(["uvicorn", "main:app", "--reload"])
-        print('checking.... prodcess.')
-        print(self.process, self.process.poll()) 
-        
+        try:
+            self.process = subprocess.Popen(["uvicorn", "main:app", "--reload"])
+            print(self.process, self.process.poll()) 
+        except Exception as e:
+            print(f"Failed to start server: {e}")
+
     def print_pid(self):
-        print('print_pid called')
         print(os.getpid())
         
     def stop_server(self):
-        print('called')
         print(os.getpid())
         if self.process:
             print(self.process, self.process.poll()) 
             print('not showing....')
-            self.process.terminate()
+            # Check if the process is still running before trying to terminate it
+            if self.process.poll() is None:
+                self.process.terminate()
             self.process = None
 
 server_manager = ServerManager()
 
 async def main():
-    print('Welcome Message')
+    print('Welcome Pitu')
     file_dir = os.path.dirname(os.path.abspath(__file__))
+    print(file_dir)
     data_dir = os.path.dirname(file_dir)
-    print('before')
     asyncio.create_task(server_manager.run_server())
-    print('eicas')
-    async for changes in awatch(os.path.join(data_dir, 'dataDir')): #cwd
-        print(changes)
+    async for changes in awatch(os.path.join(data_dir, 'dataDir')): 
+        print(f'reload: {changes}')
         asyncio.create_task(server_manager.run_server()) 
 
 if __name__ == '__main__':
@@ -55,3 +59,4 @@ if __name__ == '__main__':
         print(os.getpid())
     except asyncio.exceptions.CancelledError:
         print("Asyncio task cancelled")
+    print('With Warme...')
